+++ NEWS.md
@@ -5,14 +5,20 @@
 API:
  * DBus::Object#object_server replaces @service (which still works) and the short-lived
    @connection
+ * ObjectServer#export will raise if the path is already taken by an object
  * ObjectServer#unexport now also accepts an object path
  * Connection#object_server can export objects even without requesting any
    service name ([#49][], in beta1 already).
+ * Add PeerConnection for connections without a bus, useful for PulseAudio.
+   Fix listening for signals there ([#44][]).
+ * Moved from Connection to BusConnection: #unique_name, #proxy, #service.
+   Call send_hello in BusConnection#initialize already.
 
 Bug fixes:
  * Fixed a refactoring crasher bug in ProxyService#introspect (oops).
  * Fix crash on #unexport of /child_of_root or even /
 
+[#44]: https://github.com/mvidner/ruby-dbus/issues/44
 [#49]: https://github.com/mvidner/ruby-dbus/issues/49
 
 ## Ruby D-Bus 0.23.0.beta1 - 2023-06-05